bat commit

+++ SensingNet.v0_1/QSecs/SNetQSvidCfg.cs
@@ -12,7 +12,7 @@ namespace SensingNet.v0_1.QSecs
     /// </summary>
     public class SNetQSvidCfg
     {
-        public UInt32 QSvid;
+        public UInt64 QSvid;
         public SNetEnumStatisticsMethod StatisticsMethod = SNetEnumStatisticsMethod.Average;
         public int StatisticsSecond = 1;
 
@@ -20,7 +20,7 @@ namespace SensingNet.v0_1.QSecs
         public String DeviceIp;//可為空值
         public int DevicePort;//可為零
         public string DeviceName;//若DeviceIp/Port不存在, 以Device Name區分
-        public UInt32 DeviceSvid;
+        public UInt64 DeviceSvid;
 
         public String StoragePath;
